refactor: rework package names
