+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/organization/organizationtype/CompanyOrganizationType.java
@@ -1,4 +1,4 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.organization.definitions;
+package io.commercestacksolutions.priceproviderservice.dataaccess.organization.organizationtype;
 
 import org.springframework.stereotype.Component;
