+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/service/publicprice/strategy/PriceCandidatesQueryStrategy.java
@@ -1,7 +1,7 @@
 package io.commercestacksolutions.priceproviderservice.service.publicprice.strategy;
 
 import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.entity.PriceRowEntity;
-import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions.PriceType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.pricetype.PriceType;
 import io.commercestacksolutions.priceproviderservice.service.group.model.GroupWithDistance;
 
 import java.math.BigDecimal;
