+++ service/src/test/java/io/commercestacksolutions/priceproviderservice/service/publicprice/strategy/DefaultPriceDeterminationStrategyTest.java
@@ -2,7 +2,7 @@ package io.commercestacksolutions.priceproviderservice.service.publicprice.strat
 
 import io.commercestacksolutions.priceproviderservice.dataaccess.currency.entity.CurrencyEntity;
 import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.entity.PriceRowEntity;
-import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.definitions.PriceType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.pricetype.PriceType;
 import io.commercestacksolutions.priceproviderservice.dataaccess.taxclass.entity.TaxClassEntity;
 import io.commercestacksolutions.priceproviderservice.dataaccess.unit.entity.UnitEntity;
 import io.commercestacksolutions.priceproviderservice.service.group.model.GroupWithDistance;
